Example/Law-Firm-Website-main/Law-Firm-Website-main/requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
coverage==7.2.2
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.18
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-daterangefilter==1.0.0
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
